+++ vnstat-dashboard-core/Models/Vnstat/Total.cs
@@ -1,9 +1,11 @@
+using System;
+
 namespace vnstat_dashboard_core.Models.Vnstat{ 
 
     public class Total
     {
-        public int rx { get; set; }
-        public int tx { get; set; }
+        public Int64 rx { get; set; }
+        public Int64 tx { get; set; }
     }
 
 }
